% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Application Server Logic}
\usage{
app_server(input, output, session)
}
\description{
Defines the server-side logic for the FRDA Transcriptomic Atlas app.
}
\keyword{internal}
