% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_buildDataLong.R
\name{build_deg_long}
\alias{build_deg_long}
\title{Build long-format DEG table (dataset, gene, direction)}
\usage{
build_deg_long(
  pkg = utils::packageName(),
  p_thr = 0.05,
  lfc_min = 0,
  level = "genes"
)
}
\arguments{
\item{pkg}{Your package name (defaults to utils::packageName())}

\item{p_thr}{padj threshold to define DEGs (default 0.05)}

\item{lfc_min}{minimum |log2FC| to define DEGs (default 0)}

\item{level}{"genes" or "transcripts" (default "genes")}
}
\value{
data.frame with columns: dataset, gene, direction ∈ {Up,Down}
}
\description{
Build long-format DEG table (dataset, gene, direction)
}
