% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_deseq_loader.R
\name{deseq_loader}
\alias{deseq_loader}
\title{Loader for DESeq2 result objects saved as .rds}
\usage{
deseq_loader(dataset, level = "genes", pkg = utils::packageName())
}
\arguments{
\item{dataset}{key like "Erwin", "Mishra_FF1" (must match file stem)}

\item{level}{"genes" or "transcripts"}

\item{pkg}{package name (for system.file lookup)}
}
\value{
data.frame with columns: gene, log2FC, padj (if present), pvalue (if present)
}
\description{
Loader for DESeq2 result objects saved as .rds
}
\keyword{internal}
