% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_zenodo_data.R
\name{ensure_atlas_data}
\alias{ensure_atlas_data}
\title{Ensure FRDA atlas data are available locally (with progress bar + global banner)}
\usage{
ensure_atlas_data(keys, package = "FRDATranscriptomicAtlas")
}
\description{
Ensure FRDA atlas data are available locally (with progress bar + global banner)
}
\keyword{internal}
