% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pretty_map.R
\docType{data}
\name{pretty_map}
\alias{pretty_map}
\title{Human-readable dataset names shown in UI}
\format{
An object of class \code{character} of length 23.
}
\usage{
pretty_map
}
\description{
Human-readable dataset names shown in UI
}
\keyword{#mapping}
\keyword{datasets}
\keyword{internal}
\keyword{names}
\keyword{of}
\keyword{pretty}
\keyword{to}
