% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphTheme.R
\docType{data}
\name{theme_Marnie}
\alias{theme_Marnie}
\title{Package theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 21.
}
\usage{
theme_Marnie
}
\description{
Package theme
}
\keyword{internal}
