% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleAbout.R
\name{aboutUI}
\alias{aboutUI}
\title{About module UI}
\usage{
aboutUI(
  id,
  title = "About the FRDA Transcriptomic Atlas",
  github_url = "https://github.com/MarnieMaddock/FRDATranscriptomicAtlas",
  issues_url = "https://github.com/MarnieMaddock/FRDATranscriptomicAtlas/issues",
  contact_email = "mlm715@uowmail.edu.au"
)
}
\arguments{
\item{id}{Shiny module id}

\item{title}{Title text}

\item{github_url}{Project URL}

\item{issues_url}{Issues URL}

\item{contact_email}{Contact email}
}
\description{
About module UI
}
